separate product grid component
