--- student-store-ui/src/components/Home/Home.css
@@ -1,37 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-
-.products-grid {
-  display: inline-flex;
-  flex-wrap: wrap;
-  margin-top: 20px;
-  padding: 20px;
-  align-self: center;
-}
-
-.product-card {
-  display: inline-block;
-  vertical-align: top;
-  padding: auto;
-  margin-bottom: 30px;
-  background-color: #e4ddac;
-  /* margin-left:5%; */
-  text-align: center;
-  border-radius: 10%;
-  width: 250px;
-  display: flex;
-  flex-direction: column;
-}
-
-img {
-  width: 100%;
-  height: 200px;
-  object-fit:fill;
-}
-
-.product-info {
-  margin-top: auto;
-  font-size: 18px;
-    font-weight: 500;
-}
