+++ student-store-ui/src/components/Home/Home.jsx
@@ -3,6 +3,8 @@ import "./Home.css"
 import Hero from "../Hero/Hero";
 import SearchNFilter from "../SearchNFilter/SearchNFilter";
 import Navbar from "../Navbar/Navbar";
+import ProductGrid from "../ProductGrid/ProductGrid";
+
 
 
 import { BrowserRouter, Link } from "react-router-dom";
@@ -10,32 +12,12 @@ import { BrowserRouter, Link } from "react-router-dom";
 
 export default function Home({products}) {
 
-
-  function createProduct(info) {
-    return (
-      <section className="products-grid">
-        <div className="product-card">
-          <div className="product-img">
-            <img src={info.image} />
-          </div>
-          <div className= "product-info">
-            <p>{info.name}</p>
-            <p>${info.price}</p>
-          </div>
-        </div>
-      </section>
-    )
-    
-  }
-
   return (
     <div className="home">
       <Navbar />
       <Hero />
-      <SearchNFilter products={products}/>
-      {
-      products?.map(product => createProduct(product))
-}
+      <ProductGrid products={products} />
+      {/* <SearchNFilter products={products}/> */}
   <Navbar />
     </div>
